pay: getPaymentOptions
